Add empty prod.exs config file
config/config.exs tries to import the config file for the current Mix
environment. When including ex_aws_dynamo in an Erlang project using
rebar_mix, the Mix environment is always 'prod', so this fails because
there is no prod.exs file. Let's add an empty prod.exs file to work
around that.